Implemented a lot of features since init
Added map from google maps api with features stripped down.

Added pins and functionality to show/hide these pins.

Added sidebar with temporary hamburger icon which shows/hides the bar.
Sidebar is populated with a list of places based on the pins array.

Clicking on a pin creates an infowindow with the name of the place
and clicking on the map removes all pins.

A lot of functionality still to implement.