Many bug fixes and improvements to features
Clicking the map now closes all infowindows and resets
marker symbols.

Clicking on another marker pin now closes the previous
infowindow correctly. Additionally an infowindow will not
open when clicking on the same marker twice (preventing multiple
windows being open)

Added places from the foursquare api, with both markers and the list
of places being created from this data.

Clicking a place name now opens an infowindow and highlights the
corresponding pin.

Search feature implemented to filter by both name and restaurant
category IE 'Fenan Klein Afrika' or 'Ethiopian Restaurant' (this is
now also not case sensitive).

A center button has been added which re-centers the map if it has
moved while clicking on markers. This also closes an active infowindow
if there is one.

Continuing to improve infowindow content. This is still a work
in progress.

Added code for further info in the list view, this is not yet working
correctly and will be fixed by the next commit.