Further bug fixes and new features
Clicking on an item in the sidebar will now show further info
correctly from within the location.

Improved the code for clicking between markers and opening/closing
infowindows. Preivously state was being mutated directly, which is
incorrect. This now works properly.

Clicking the close button on the infowindow now resets the marker.
This means that you can click the same marker again to reopen the
infowindow. This was previously not working as intended.